+++ packages/code-analyzer/src/handlers/handleConstructors.ts
@@ -1,0 +1,27 @@
+import { ClassDeclaration, Scope } from 'ts-morph';
+import { Entry } from '../models/entry';
+import { Kind } from '../models/kind';
+import { findNonTrivialParameters, handleParameters } from './handleParameters';
+
+export function handleConstructors(cls: ClassDeclaration): Entry[] {
+    const constructors: Entry[] = [];
+    cls.getConstructors().map(c => {
+        const scope = c.getScope();
+
+        if (scope === Scope.Private || scope === Scope.Protected) {
+            return;
+        }
+
+        constructors.push({
+            kind: Kind[Kind.Constructor],
+            name: 'constructor',
+            docs: c.getJsDocs()[0]?.getText(),
+            parent: cls.getName(),
+            parentKind: Kind[Kind.Class],
+            dependencies: findNonTrivialParameters(c.getParameters()),
+            parameters: handleParameters(c.getParameters()),
+        });
+    });
+
+    return constructors;
+}
