+++ packages/code-analyzer/src/models/ingestionRequest.ts
@@ -1,4 +1,4 @@
-import { Entry } from "./entry"
+import { Entry } from './entry';
 
 export type IngestionRequest = {
     entries: Entry[];
