+++ packages/code-analyzer/src/models/kind.ts
