+++ packages/code-analyzer/src/models/property.ts
